+++ .gitignore
@@ -137,3 +137,14 @@ dist
 
 # Electron Forge build output
 out/
+
+# ESLint
+.eslintcache
+
+# IDE - VSCode
+.vscode/*
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
+.history
