+++ CHANGELOG.md
@@ -2,21 +2,18 @@
 
 ## [1.2.0](https://github.com/artmakh/dxvk-version-mananger/compare/v1.1.0...v1.2.0) (2025-03-25)
 
-
 ### Features
 
-* Set d3d and win exec version ([#10](https://github.com/artmakh/dxvk-version-mananger/issues/10)) ([943167a](https://github.com/artmakh/dxvk-version-mananger/commit/943167a2276904f7d21b6a216abf9642da52120c))
+- Set d3d and win exec version ([#10](https://github.com/artmakh/dxvk-version-mananger/issues/10)) ([943167a](https://github.com/artmakh/dxvk-version-mananger/commit/943167a2276904f7d21b6a216abf9642da52120c))
 
 ## [1.1.0](https://github.com/artmakh/dxvk-version-mananger/compare/v1.0.0...v1.1.0) (2025-03-25)
 
-
 ### Features
 
-* Delete DXVK on non-backuped games ([#7](https://github.com/artmakh/dxvk-version-mananger/issues/7)) ([a208779](https://github.com/artmakh/dxvk-version-mananger/commit/a2087792fc9b10931b6400f2674fe27c237fb33f))
+- Delete DXVK on non-backuped games ([#7](https://github.com/artmakh/dxvk-version-mananger/issues/7)) ([a208779](https://github.com/artmakh/dxvk-version-mananger/commit/a2087792fc9b10931b6400f2674fe27c237fb33f))
 
 ## 1.0.0 (2025-03-23)
 
-
 ### Features
 
-* First release ([18f2065](https://github.com/artmakh/dxvk-version-mananger/commit/18f20658c54c57243813f57d8037932aa3d79bb4))
+- First release ([18f2065](https://github.com/artmakh/dxvk-version-mananger/commit/18f20658c54c57243813f57d8037932aa3d79bb4))
