+++ README.md
@@ -6,7 +6,7 @@
 
 A Windows utility to manage versions of DXVK in installed games.
 
-![](assets/screen.gif)
+![](.github/screen.gif)
 
 ## Features
 
@@ -83,3 +83,32 @@ This project uses [Conventional Commits](https://www.conventionalcommits.org/) s
 ## License
 
 This project is licensed under the MIT License - see the [LICENSE](LICENSE) file for details.
+
+## Development
+
+### Project Structure
+
+The project follows a standard Electron application structure:
+
+```
+dxvk-version-manager/
+├── src/
+│   ├── main/             # Main process files
+│   ├── renderer/         # Renderer process files
+│   │   ├── assets/       # Images and other assets
+│   │   ├── styles/       # CSS files
+│   ├── preload/          # Preload scripts
+│   └── common/           # Shared code between processes
+├── .eslintrc.json        # ESLint configuration
+├── .prettierrc           # Prettier configuration
+└── package.json          # Project metadata and dependencies
+```
+
+### Development Scripts
+
+- `npm start` - Start the application in development mode
+- `npm run lint` - Run ESLint to check code quality
+- `npm run lint:fix` - Run ESLint and automatically fix issues
+- `npm run format` - Run Prettier to format code
+- `npm run package` - Package the application for distribution
+- `npm run make` - Create distribution packages for the application
