+++ package.json
@@ -2,14 +2,16 @@
   "name": "dxvk-version-mananger",
   "version": "1.2.0",
   "description": "Windows utility to manage versions of DXVK in installed games",
-  "main": "main.js",
+  "main": "src/main/index.js",
   "scripts": {
     "start": "electron .",
     "start:forge": "electron-forge start",
     "package": "electron-forge package",
     "make": "electron-forge make",
     "publish": "electron-forge publish",
-    "lint": "echo \"No linting configured\""
+    "lint": "eslint --ext .js . --config .eslintrc.json",
+    "lint:fix": "eslint --ext .js . --fix --config .eslintrc.json",
+    "format": "prettier --write \"**/*.{js,jsx,json,md}\""
   },
   "repository": {
     "type": "git",
@@ -28,7 +30,13 @@
   "devDependencies": {
     "@electron-forge/cli": "^7.2.0",
     "@electron-forge/maker-zip": "^7.2.0",
-    "electron": "^35.0.3"
+    "@eslint/js": "^9.23.0",
+    "electron": "^35.0.3",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^10.1.1",
+    "eslint-plugin-prettier": "^5.2.5",
+    "globals": "^16.0.0",
+    "prettier": "^3.5.3"
   },
   "dependencies": {
     "electron-squirrel-startup": "^1.0.0",
@@ -40,7 +48,11 @@
         "asar": true,
         "name": "DXVK Version Manager",
         "platform": "win32",
-        "arch": "x64"
+        "arch": "x64",
+        "icon": "src/renderer/assets/icon.ico",
+        "extraResource": [
+          "src/renderer/assets"
+        ]
       },
       "makers": [
         {
