+++ src/renderer/assets/placeholder.png
