+++ src/renderer/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>DXVK Version Manager</title>
-    <link rel="stylesheet" href="styles.css">
+    <link rel="stylesheet" href="styles/index.css">
 </head>
 <body>
     <div class="container">
@@ -40,6 +40,6 @@
         </div>
     </div>
 
-    <script src="renderer.js"></script>
+    <script src="index.js"></script>
 </body>
 </html>
